+++ Ladybug/App/AppDelegate.swift
@@ -17,6 +17,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
 
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let rootViewController = TabBarController()
+        window?.rootViewController = rootViewController
+        window?.makeKeyAndVisible()
+
+        UIView.appearance().tintColor = .tintColor
+
         return true
     }