+++ Ladybug/Extensions/UIColor+Extension.swift
@@ -10,4 +10,5 @@ import UIKit
 
 extension UIColor {
     static let barTintColor: UIColor = UIColor.black
+    static let tintColor: UIColor = #colorLiteral(red: 0.9450980392, green: 0.2352941176, blue: 0.3215686275, alpha: 1)
 }