+++ Ladybug/Extensions/UINavigationController+Extension.swift
@@ -1,0 +1,15 @@
+//
+//  UINavigationController+Extension.swift
+//  Ladybug
+//
+//  Created by Ethanhuang on 2018/6/22.
+//  Copyright © 2018年 Elaborapp Co., Ltd. All rights reserved.
+//
+
+import UIKit
+
+extension UINavigationController {
+    override open var preferredStatusBarStyle: UIStatusBarStyle {
+        return .lightContent
+    }
+}