+++ Ladybug/Recent/RecentViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import SafariServices
 
 class RecentViewController: UITableViewController {
 
@@ -15,45 +14,13 @@ class RecentViewController: UITableViewController {
         super.viewDidLoad()
 
         navigationItem.title = "Recent".localized()
-
-        RadarURLOpener.shared.delegate = self
+        if #available(iOS 11.0, *) {
+            navigationController?.navigationBar.prefersLargeTitles = true
+        }
     }
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
-    }
-}
-
-extension RecentViewController: RadarURLOpenerUI {
-    func openRadarInSafariViewController(_ radarID: RadarID, radarOption: RadarOption, readerMode: Bool) {
-        // TODO: Prepend radarID to self.array
-
-        let url = radarID.url(by: radarOption)
-        presentSafariViewController(url: url, readerMode: readerMode)
-    }
-
-    private func presentSafariViewController(url: URL, readerMode: Bool) {
-        let sfvc: SFSafariViewController = {
-            if #available(iOS 11.0, *) {
-                let config = SFSafariViewController.Configuration()
-                config.barCollapsingEnabled = false
-                config.entersReaderIfAvailable = readerMode
-
-                return SFSafariViewController(url: url, configuration: config)
-            } else {
-                return SFSafariViewController(url: url, entersReaderIfAvailable: readerMode)
-            }
-        }()
-
-        sfvc.preferredBarTintColor = .barTintColor
-
-        if let presented = self.presentedViewController {
-            presented.dismiss(animated: false) {
-                self.present(sfvc, animated: false, completion: nil)
-            }
-        } else {
-            self.present(sfvc, animated: false, completion: nil)
-        }
     }
 }