+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.1(1)] - 2018-06-23
 ### Added
 - Radar model, RadarURLParser, RadarURLBuilder protocols, implementation, and tests
 - Basic UI architecture
@@ -13,3 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Present in SFSafariViewController
 - Use Main.storyboard as launch screen file
 - RadarOption and BrowserOption in settings
+- App Store, developer's Twitter, and GitHub links
+- Feedback email composer
+- App icon
+- Theme